requirements.txt
uAgents
pydantic
requests
python-dotenv
email-validator
